celery: rename celery.py into workers.py to avoid all possible nameclashing with the Celery dependency module
